# dataset get-rows and row-maps optimized