@mikera 2015-04-12
Looks like a pretty good improvement! Thanks!